Added additional content
EDFDB 1 row: is a file that you can run in MySQL that inserts 1 line of data into each row for all of the tables. 
edfDB Query Test: are a bunch of queries I created to run different reports on the data. We may be able to use this later if we setup some reports. 
edfDB Query: This is the original database for our form. 
EDFForm Teamb: This is the form that associates all of the column names to the user input.